Use Oj correctly as the JSON engine
Though Oj was originally intended to be used (hence it was in the
Gemfile), it was not being used by all of Rabl, AMS, and presenters.

In fact, it was only being used by Rabl as it is the engine Rabl
uses by default. In othewords, the benchmark was unfair to AMS
and presenters. With the Oj boost to AMS and presenters, the
relative performance of Rabl gets even worse.